Refactor numbers with mixins
