+++ include/xwidgets/xnumber_bounded.hpp
@@ -11,9 +11,10 @@
 
 #include <type_traits>
 
+#include "mixin/xnumeric.hpp"
 #include "xdescription_style.hpp"
 #include "xmaterialize.hpp"
-#include "xnumber_impl.hpp"
+#include "xwidget.hpp"
 
 namespace xw
 {
@@ -22,13 +23,15 @@ namespace xw
      *********************************/
 
     template <class D>
-    class xnumber_bounded : public xnumber_bounded_impl<D>
+    class xnumber_bounded : public xwidget<D>,
+                            public mixin::xnumeric_bounded<D>
     {
     public:
 
-        using base_type = xnumber_bounded_impl<D>;
         using derived_type = D;
-        using typename base_type::value_type;
+        using base_type = xwidget<D>;
+        using mixin_numeric_type = mixin::xnumeric_bounded<D>;
+        using typename mixin_numeric_type::value_type;
 
         void serialize_state(nl::json&, xeus::buffer_sequence&) const;
         void apply_patch(const nl::json&, const xeus::buffer_sequence&);
@@ -66,6 +69,7 @@ namespace xw
     inline void xnumber_bounded<D>::serialize_state(nl::json& state, xeus::buffer_sequence& buffers) const
     {
         base_type::serialize_state(state, buffers);
+        mixin_numeric_type::serialize_state(state, buffers);
 
         xwidgets_serialize(continuous_update(), state["continuous_update"], buffers);
         xwidgets_serialize(description(), state["description"], buffers);
@@ -79,6 +83,7 @@ namespace xw
     inline void xnumber_bounded<D>::apply_patch(const nl::json& patch, const xeus::buffer_sequence& buffers)
     {
         base_type::apply_patch(patch, buffers);
+        mixin_numeric_type::apply_patch(patch, buffers);
 
         set_property_from_patch(continuous_update, patch, buffers);
         set_property_from_patch(description, patch, buffers);
