+++ include/xwidgets/xplay.hpp
@@ -9,9 +9,10 @@
 #ifndef XWIDGETS_PLAY_HPP
 #define XWIDGETS_PLAY_HPP
 
+#include "mixin/xnumeric.hpp"
 #include "xdescription_style.hpp"
 #include "xmaterialize.hpp"
-#include "xnumber_impl.hpp"
+#include "xwidget.hpp"
 
 namespace xw
 {
@@ -20,13 +21,15 @@ namespace xw
      ********************/
 
     template <class D>
-    class xplay : public xnumber_bounded_impl<D>
+    class xplay : public xwidget<D>,
+                  public mixin::xnumeric_bounded<D>
     {
     public:
 
-        using base_type = xnumber_bounded_impl<D>;
         using derived_type = D;
-        using typename base_type::value_type;
+        using base_type = xwidget<D>;
+        using mixin_numeric_type = mixin::xnumeric_bounded<D>;
+        using typename mixin_numeric_type::value_type;
 
         void serialize_state(nl::json& state, xeus::buffer_sequence&) const;
         void apply_patch(const nl::json&, const xeus::buffer_sequence&);
@@ -66,6 +69,7 @@ namespace xw
     inline void xplay<D>::serialize_state(nl::json& state, xeus::buffer_sequence& buffers) const
     {
         base_type::serialize_state(state, buffers);
+        mixin_numeric_type::serialize_state(state, buffers);
 
         xwidgets_serialize(description(), state["description"], buffers);
         xwidgets_serialize(description_allow_html(), state["description_allow_html"], buffers);
@@ -82,6 +86,7 @@ namespace xw
     inline void xplay<D>::apply_patch(const nl::json& patch, const xeus::buffer_sequence& buffers)
     {
         base_type::apply_patch(patch, buffers);
+        mixin_numeric_type::apply_patch(patch, buffers);
 
         set_property_from_patch(description, patch, buffers);
         set_property_from_patch(description_allow_html, patch, buffers);
@@ -96,7 +101,6 @@ namespace xw
 
     template <class D>
     inline xplay<D>::xplay()
-        : base_type()
     {
         set_defaults();
     }
