+++ include/xwidgets/xprogress.hpp
@@ -14,11 +14,12 @@
 
 #include <xtl/xoptional.hpp>
 
+#include "mixin/xnumeric.hpp"
 #include "xcolor.hpp"
 #include "xeither.hpp"
 #include "xmaterialize.hpp"
-#include "xnumber_impl.hpp"
 #include "xstyle.hpp"
+#include "xwidget.hpp"
 
 namespace xw
 {
@@ -56,13 +57,15 @@ namespace xw
      ***********************/
 
     template <class D>
-    class xprogress : public xnumber_bounded_impl<D>
+    class xprogress : public xwidget<D>,
+                      public mixin::xnumeric_bounded<D>
     {
     public:
 
-        using base_type = xnumber_bounded_impl<D>;
         using derived_type = D;
-        using typename base_type::value_type;
+        using base_type = xwidget<D>;
+        using mixin_numeric_type = mixin::xnumeric_bounded<D>;
+        using typename mixin_numeric_type::value_type;
 
         void serialize_state(nl::json&, xeus::buffer_sequence&) const;
         void apply_patch(const nl::json&, const xeus::buffer_sequence&);
@@ -135,6 +138,7 @@ namespace xw
     inline void xprogress<D>::serialize_state(nl::json& state, xeus::buffer_sequence& buffers) const
     {
         base_type::serialize_state(state, buffers);
+        mixin_numeric_type::serialize_state(state, buffers);
 
         xwidgets_serialize(bar_style(), state["bar_style"], buffers);
         xwidgets_serialize(description(), state["description"], buffers);
@@ -147,6 +151,7 @@ namespace xw
     inline void xprogress<D>::apply_patch(const nl::json& patch, const xeus::buffer_sequence& buffers)
     {
         base_type::apply_patch(patch, buffers);
+        mixin_numeric_type::apply_patch(patch, buffers);
 
         set_property_from_patch(bar_style, patch, buffers);
         set_property_from_patch(description, patch, buffers);
